+++ probe-rs-target/src/chip_family.rs
@@ -118,16 +118,21 @@ pub struct ChipFamily {
     /// E.g. `nRF52832`.
     pub name: String,
     /// The JEP106 code of the manufacturer.
-    #[cfg_attr(
-        not(feature = "bincode"),
-        serde(skip_serializing_if = "Option::is_none")
-    )]
     pub manufacturer: Option<JEP106Code>,
+    /// The `target-gen` process will set this to `true`.
+    /// Please change this to `false` if this file is modified from the generated, or is a manually created target description.
+    #[serde(default)]
+    pub generated_from_pack: bool,
+    /// The latest release of the pack file from which this was generated.
+    /// Values:
+    /// - `Some("1.3.0")` if the latest pack file release was for example "1.3.0".
+    /// - `None` if this was not generated from a pack file, or has been modified since it was generated.
+    #[serde(default)]
+    pub pack_file_release: Option<String>,
     /// This vector holds all the variants of the family.
     pub variants: Vec<Chip>,
     /// This vector holds all available algorithms.
     pub flash_algorithms: Vec<RawFlashAlgorithm>,
-
     #[serde(skip, default = "default_source")]
     /// Source of the target description, used for diagnostics
     pub source: TargetDescriptionSource,