+++ probe-rs/src/config/registry.rs
@@ -45,6 +45,8 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
         ChipFamily {
             name: "Generic ARMv6-M".to_owned(),
             manufacturer: None,
+            generated_from_pack: false,
+            pack_file_release: None,
             variants: vec![
                 Chip::generic_arm("Cortex-M0", CoreType::Armv6m),
                 Chip::generic_arm("Cortex-M0+", CoreType::Armv6m),
@@ -57,6 +59,8 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
         ChipFamily {
             name: "Generic ARMv7-M".to_owned(),
             manufacturer: None,
+            generated_from_pack: false,
+            pack_file_release: None,
             variants: vec![Chip::generic_arm("Cortex-M3", CoreType::Armv7m)],
             flash_algorithms: vec![],
             source: TargetDescriptionSource::Generic,
@@ -64,6 +68,8 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
         ChipFamily {
             name: "Generic ARMv7E-M".to_owned(),
             manufacturer: None,
+            generated_from_pack: false,
+            pack_file_release: None,
             variants: vec![
                 Chip::generic_arm("Cortex-M4", CoreType::Armv7em),
                 Chip::generic_arm("Cortex-M7", CoreType::Armv7em),
@@ -74,6 +80,8 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
         ChipFamily {
             name: "Generic ARMv8-M".to_owned(),
             manufacturer: None,
+            generated_from_pack: false,
+            pack_file_release: None,
             variants: vec![
                 Chip::generic_arm("Cortex-M23", CoreType::Armv8m),
                 Chip::generic_arm("Cortex-M33", CoreType::Armv8m),
@@ -86,6 +94,8 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
         ChipFamily {
             name: "Generic RISC-V".to_owned(),
             manufacturer: None,
+            pack_file_release: None,
+            generated_from_pack: false,
             variants: vec![Chip {
                 name: "riscv".to_owned(),
                 part: None,
@@ -96,6 +106,7 @@ fn add_generic_targets(vec: &mut Vec<ChipFamily>) {
                 }],
                 memory_map: vec![],
                 flash_algorithms: vec![],
+                supports_connect_under_reset: false,
             }],
             flash_algorithms: vec![],
             source: TargetDescriptionSource::Generic,
@@ -114,8 +125,13 @@ impl Registry {
     fn from_builtin_families() -> Self {
         const BUILTIN_TARGETS: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/targets.bincode"));
 
-        let mut families: Vec<ChipFamily> = bincode::deserialize(BUILTIN_TARGETS)
-            .expect("Failed to deserialize builtin targets. This is a bug.");
+        let mut families: Vec<ChipFamily> = match bincode::deserialize(BUILTIN_TARGETS) {
+            Ok(families) => families,
+            Err(err) => panic!(
+                "Failed to deserialize builtin targets. This is a bug : {:?}",
+                err
+            ),
+        };
 
         add_generic_targets(&mut families);