+++ probe-rs/src/config/target.rs
@@ -29,12 +29,12 @@ pub struct Target {
     pub flash_algorithms: Vec<RawFlashAlgorithm>,
     /// The memory map of the target.
     pub memory_map: Vec<MemoryRegion>,
-
     /// Source of the target description. Used for diagnostics.
     pub(crate) source: TargetDescriptionSource,
-
     /// Debug sequences for the given target.
     pub debug_sequence: DebugSequence,
+    /// This target supports the `connect_under_reset` feature (see [`probe_rs::probe::Probe::attach_under_reset`]).
+    pub supports_connect_under_reset: bool,
 }
 
 impl std::fmt::Debug for Target {
@@ -135,6 +135,9 @@ impl Target {
             source: family.source.clone(),
             memory_map: chip.memory_map.clone(),
             debug_sequence,
+            // Note: This is a solution to the fact that not ALL STM32 target files have been updated to generate this field.
+            supports_connect_under_reset: chip.supports_connect_under_reset
+                || chip.name.starts_with("STM32"),
         })
     }