+++ target-gen/src/main.rs
@@ -4,14 +4,9 @@ pub mod flash_device;
 pub mod generate;
 pub mod parser;
 
-use std::{
-    fs::{create_dir, File, OpenOptions},
-    io::Write,
-    path::{Path, PathBuf},
-};
-
 use anyhow::{bail, ensure, Context, Result};
 use clap::Parser;
+use parser::extract_flash_algo;
 use probe_rs::{
     config::{
         Chip, ChipFamily, Core, MemoryRegion, NvmRegion, RamRegion,
@@ -21,8 +16,12 @@ use probe_rs::{
 };
 use probe_rs_target::{ArmCoreAccessOptions, CoreAccessOptions};
 use simplelog::*;
-
-use parser::extract_flash_algo;
+use std::{
+    env::current_dir,
+    fs::{create_dir, File, OpenOptions},
+    io::{BufRead, Write},
+    path::{Path, PathBuf},
+};
 
 #[derive(clap::Parser)]
 enum TargetGen {
@@ -41,14 +40,30 @@ enum TargetGen {
         )]
         output_dir: PathBuf,
     },
-    /// Generates all the target descriptions from the entries listed in the ARM root VIDX/PIDX at <https://www.keil.com/pack/Keil.pidx>.
+    /// Generates from the entries listed in the ARM root VIDX/PIDX at <https://www.keil.com/pack/Keil.pidx>.
+    /// This will only download and generate target descriptions for chip families that are already supported by probe-rs, to avoid generating a lot of unsupportable chip families.
+    /// Please use the `pack` subcommand to generate target descriptions for other chip families.
     Arm {
+        /// Only download and generate target descriptions for the chip families that start with the specified name, e.g. `STM32H7` or `LPC55S69`.
+        #[clap(
+            long = "list",
+            short = 'l',
+            help = "Optionally, list the names of all pack files available in <https://www.keil.com/pack/Keil.pidx>"
+        )]
+        list: bool,
+        /// Only download and generate target descriptions for the pack files that start with the specified name`.
+        #[clap(
+            long = "filter",
+            short = 'f',
+            help = "Optionally, filter the pack files that start with the specified name,\ne.g. `STM32H7xx` or `LPC55S69_DFP`.\nSee `target-gen arm --list` for a list of available Pack files"
+        )]
+        pack_filter: Option<String>,
         #[clap(
             name = "OUTPUT",
             value_parser,
             help = "An output directory where all the generated .yaml files are put in."
         )]
-        output_dir: PathBuf,
+        output_dir: Option<PathBuf>,
     },
     /// Extract a flash algorithm from an ELF file
     Elf {
@@ -90,7 +105,11 @@ fn main() -> Result<()> {
             update,
             name,
         } => cmd_elf(elf, output, update, name)?,
-        TargetGen::Arm { output_dir } => cmd_arm(output_dir.as_path())?,
+        TargetGen::Arm {
+            output_dir,
+            pack_filter: chip_family,
+            list,
+        } => cmd_arm(output_dir, chip_family, list)?,
     }
 
     println!("Finished in {:?}", t.elapsed());
@@ -149,8 +168,7 @@ fn cmd_elf(
         }
 
         let target_description = File::create(&target_description_file)?;
-
-        serde_yaml::to_writer(&target_description, &family)?;
+        serialize_to_yaml_file(&family, &target_description)?;
     } else {
         // Create a complete target specification, with place holder values
         let algorithm_name = algorithm.name.clone();
@@ -159,6 +177,8 @@ fn cmd_elf(
         let chip_family = ChipFamily {
             name: "<family name>".to_owned(),
             manufacturer: None,
+            generated_from_pack: false,
+            pack_file_release: None,
             variants: vec![Chip {
                 cores: vec![Core {
                     name: "main".to_owned(),
@@ -187,17 +207,16 @@ fn cmd_elf(
                     }),
                 ],
                 flash_algorithms: vec![algorithm_name],
+                supports_connect_under_reset: false,
             }],
             flash_algorithms: vec![algorithm],
             source: BuiltIn,
         };
 
-        let serialized = serde_yaml::to_string(&chip_family)?;
-
         match output {
             Some(output) => {
                 // Ensure we don't overwrite an existing file
-                let mut file = OpenOptions::new()
+                let file = OpenOptions::new()
                     .write(true)
                     .create_new(true)
                     .open(&output)
@@ -205,10 +224,9 @@ fn cmd_elf(
                         "Failed to create target file '{}'.",
                         output.display()
                     ))?;
-
-                file.write_all(serialized.as_bytes())?;
+                serialize_to_yaml_file(&chip_family, &file)?;
             }
-            None => println!("{}", serialized),
+            None => println!("{}", serde_yaml::to_string(&chip_family)?),
         }
     }
 
@@ -253,10 +271,11 @@ fn cmd_pack(input: &Path, out_dir: &Path) -> Result<()> {
     let mut generated_files = Vec::with_capacity(families.len());
 
     for family in &families {
-        let path = out_dir.join(family.name.clone() + ".yaml");
+        let path = out_dir.join(family.name.clone().replace(' ', "_") + ".yaml");
         let file = std::fs::File::create(&path)
             .context(format!("Failed to create file '{}'.", path.display()))?;
-        serde_yaml::to_writer(file, &family)?;
+
+        serialize_to_yaml_file(family, &file)?;
 
         generated_files.push(path);
     }
@@ -272,7 +291,24 @@ fn cmd_pack(input: &Path, out_dir: &Path) -> Result<()> {
 
 /// Handle the arm subcommand.
 /// Generated target descriptions will be placed in `out_dir`.
-fn cmd_arm(out_dir: &Path) -> Result<()> {
+fn cmd_arm(out_dir: Option<PathBuf>, chip_family: Option<String>, list: bool) -> Result<()> {
+    if list {
+        let mut packs = crate::fetch::get_vidx()?;
+        println!("Available ARM CMSIS Pack files:");
+        packs.pdsc_index.sort_by(|a, b| a.name.cmp(&b.name));
+        for pack in packs.pdsc_index.iter() {
+            println!("\t{}", pack.name);
+        }
+        return Ok(());
+    }
+
+    let out_dir = if let Some(target_dir) = out_dir {
+        target_dir.as_path().to_owned()
+    } else {
+        log::info!("No output directory specified. Using current directory.");
+        current_dir()?
+    };
+
     if !out_dir.exists() {
         create_dir(&out_dir).context(format!(
             "Failed to create output directory '{}'.",
@@ -282,15 +318,15 @@ fn cmd_arm(out_dir: &Path) -> Result<()> {
 
     let mut families = Vec::<ChipFamily>::new();
 
-    generate::visit_arm_files(&mut families)?;
+    generate::visit_arm_files(&mut families, chip_family)?;
 
     let mut generated_files = Vec::with_capacity(families.len());
 
     for family in &families {
-        let path = out_dir.join(family.name.clone() + ".yaml");
+        let path = out_dir.join(family.name.clone().replace(' ', "_") + ".yaml");
         let file = std::fs::File::create(&path)
             .context(format!("Failed to create file '{}'.", path.display()))?;
-        serde_yaml::to_writer(file, &family)?;
+        serialize_to_yaml_file(family, &file)?;
 
         generated_files.push(path);
     }
@@ -301,5 +337,34 @@ fn cmd_arm(out_dir: &Path) -> Result<()> {
         println!("\t{}", file.display());
     }
 
+    Ok(())
+}
+
+/// Some optimizations to improve the readability of the `serde_yaml` output:
+/// - If `Option<T>` is `None`, it is serialized as `null` ... we want to omit it.
+/// - If `Vec<T>` is empty, it is serialized as `[]` ... we want to omit it.
+/// - `serde_yaml` serializes hex formatted integers as single quoted strings, e.g. '0x1234' ... we need to remove the single quotes so that it round-trips properly.
+fn serialize_to_yaml_file(family: &ChipFamily, file: &File) -> Result<(), anyhow::Error> {
+    let yaml_string = serde_yaml::to_string(&family)?;
+    let mut reader = std::io::BufReader::new(yaml_string.as_bytes());
+    let mut reader_line = String::new();
+    let mut writer = std::io::BufWriter::new(file);
+    while reader.read_line(&mut reader_line)? > 0 {
+        if reader_line.ends_with(": null\n")
+            || reader_line.ends_with(": []\n")
+            || reader_line.ends_with(": false\n")
+        {
+            // Skip the line
+        } else if (reader_line.contains("'0x") || reader_line.contains("'0X"))
+            && reader_line.ends_with("'\n")
+        {
+            // Remove the single quotes
+            reader_line = reader_line.replace('\'', "");
+            writer.write_all(reader_line.as_bytes())?;
+        } else {
+            writer.write_all(reader_line.as_bytes())?;
+        }
+        reader_line.clear();
+    }
     Ok(())
 }